requirements.txt
numpy>=1.13.1
pandas>=0.19
biopython>=1.69
pycf>=0.6.8
